chore(cli): move key related cli to key namespace
